Upgraded Milestones view to Angular component. Also removed code and template for allowing teachers to create milestone based on completion. #2850
