+++ src/main/webapp/site/src/app/teacher-hybrid-angular.module.ts
@@ -4,6 +4,7 @@ import '../../../wise5/teacher/teacher-angular-js-module';
 import { UpgradeModule } from '@angular/upgrade/static';
 import { setUpLocationSync } from '@angular/router/upgrade';
 import { ProjectService } from '../../../wise5/services/projectService';
+import { MilestonesComponent } from './classroom-monitor/milestones/milestones.component';
 import { MilestoneReportDataComponent } from './teacher/milestone/milestone-report-data/milestone-report-data.component';
 import { TeacherProjectService } from '../../../wise5/services/teacherProjectService';
 import { ProjectAssetService } from './services/projectAssetService';
@@ -53,6 +54,7 @@ import { NavItemProgressComponent } from './classroom-monitor/nav-item-progress/
     EditComponentTagsComponent,
     EditComponentWidthComponent,
     ManageStudentsComponent,
+    MilestonesComponent,
     MilestoneReportDataComponent,
     NavItemProgressComponent,
     NodeAdvancedGeneralAuthoringComponent,
