--- src/main/webapp/wise5/classroomMonitor/milestones/milestones.html
@@ -1,43 +1,0 @@
-<div class="view-content view-content--with-sidemenu">
-    <div class="l-constrained">
-        <div layout="row" layout-align="center center" layout-align-gt-xs="start center" layout-wrap>
-            <!-- TODO: make into MilestoneItem component -->
-            <md-card ng-repeat="milestone in milestonesController.milestones | orderBy:'params.targetDateTemp'"
-                     class="milestone md-button" md-ink-ripple
-                     ng-click="milestonesController.showMilestoneDetails(milestone, $event)"
-                     aria-label="{{ ::'VIEW_MILESTONE_DETAILS' | translate:{ name: milestone.name } }}">
-                <md-card-content layout="column" layout-align="center center">
-                    <div class="milestone__title" md-truncate>{{ milestone.name }}</div>
-                    <div class="milestone__progress">
-                        <md-progress-circular md-diameter="96"
-                                              md-mode="determinate"
-                                              value="{{ milestone.percentageCompleted }}"
-                                              ng-class="{ 'success': milestone.percentageCompleted === 100 }"></md-progress-circular>
-                        <div class="milestone__progress__percent"
-                             layout="row" layout-align="center center">
-                            <span ng-if="milestone.percentageCompleted < 100">{{ milestone.percentageCompleted }}%</span>
-                            <md-icon ng-if="milestone.percentageCompleted === 100" class="success md-48"> check </md-icon>
-                        </div>
-                    </div>
-                    <p class="md-body-1 text-secondary">{{ milestone.numberOfStudentsCompleted }}/{{ milestone.numberOfStudentsInRun }} teams completed</p>
-                    <p ng-if="milestone.isReportAvailable" class="md-body-2 info">~ {{ ::'reportAvailable' | translate }} ~</p>
-                    <div class="md-body-2" ng-if="!milestone.isReportAvailable"
-                         ng-class="{ 'warn': milestonesController.isBeforeDay(milestone.params.targetDate, milestone.percentageCompleted),
-                            'info': milestonesController.isSameDay(milestone.params.targetDate, milestone.percentageCompleted) }">
-                        {{ milestone.params.targetDate | amCalendar }}
-                    </div>
-                </md-card-content>
-            </md-card>
-            <!-- <md-card class="milestone milestone--add md-button primary" md-ink-ripple
-                     ng-click="milestonesController.editMilestone(null, $event)"
-                     aria-label="{{ ::'ADD_MILESTONE' | translate }}">
-                <md-card-content layout="column" layout-align="center center">
-                    <div class="milestone__title center">{{ ::'ADD_MILESTONE' | translate }}</div>
-                    <div class="milestone--add__icon" layout="column" layout-align="center center">
-                        <md-icon class="md-48 md-primary"> add </md-icon>
-                    </div>
-                </md-card-content>
-            </md-card> -->
-        </div>
-    </div>
-</div>
