+++ src/main/webapp/wise5/i18n/i18n_ar.json
@@ -1,6 +1,5 @@
 {
   "ADD_CONNECTED_COMPONENT": "إضافة مكون متصل",
-  "ADD_MILESTONE": "إضافة حدث رئيسي",
   "ADD_TO_NOTEBOOK": "إضافة إلى دفتر ملاحظات",
   "ADVANCED": "متقدم",
   "ALERTS": "تنبيهات",
@@ -18,15 +17,12 @@
   "CONNECTED_COMPONENTS": "المكونات المتصلة",
   "CORRECT": "صحيح",
   "DELETE": "حذف",
-  "DELETE_MILESTONE": "حذف الحدث الرئيسي ",
-  "DELETE_MILESTONE_CONFIRM": "هل أنت متأكد من رغبتك في حذف هذا الحدث الرئيسي بصورة دائمة؟",
   "DESCRIPTION": "التفاصيل",
   "DETAILS": "التفاصيل",
   "DONE": "تم",
   "DOWN": "لأسفل",
   "EDIT": "التحرير",
   "EDIT_COMPONENT_RUBRIC": "تحرير معايير المكون",
-  "EDIT_MILESTONE": "تحرير المرحلة",
   "ENTER_PROMPT_HERE": "أدخل مطالبة هنا",
   "ERROR_CHECK_YOUR_INTERNET_CONNECTION": " خطأ: لا يجري حفظ البيانات! تحقق من اتصال الإنترنت لديك.",
   "ERROR_COULD_NOT_CALCULATE_IS_COMPLETED": "خطأ:لم يتم الحساب  تم الانتهاء () من أجل   \nمكون بمعرف",
