+++ src/main/webapp/wise5/i18n/i18n_el.json
@@ -1,6 +1,5 @@
 {
   "ADD_CONNECTED_COMPONENT": "Πρόσθεση Συνδεδεμένου Στοιχείου",
-  "ADD_MILESTONE": "Πρόσθεση Ορόσημου",
   "ADD_TO_NOTEBOOK": "Πρόσθεση στο Σημειωματάριο",
   "ADVANCED": "Για Προχωρημένους",
   "ALERTS": "Ειδοποιήσεις",
@@ -16,15 +15,12 @@
   "COMPONENT_WIDTH": "Πλάτος Στοιχείου",
   "CORRECT": "Σωστό",
   "DELETE": "Διαγραφή",
-  "DELETE_MILESTONE": "Διαγραφή Ορόσημου",
-  "DELETE_MILESTONE_CONFIRM": "Θέλετε σίγουρα να διαγράψετε μόνιμα αυτό το ορόσημο;",
   "DESCRIPTION": "Περιγραφή",
   "DETAILS": "Λεπτομέρειες",
   "DONE": "Έτοιμο",
   "DOWN": "Κάτω",
   "EDIT": "Επεξεργασία",
   "EDIT_COMPONENT_RUBRIC": "Επεξεργασία Οδηγού Στοιχείου",
-  "EDIT_MILESTONE": "Επεξεργασία Ορόσημου",
   "ENTER_PROMPT_HERE": "Εισαγωγή Υπόδειξης",
   "ERROR_CHECK_YOUR_INTERNET_CONNECTION": "Σφάλμα: τα δεδομένα δεν αποθηκεύονται! Ελέγξτε τη σύνδεσή σας στο internet.",
   "ERROR_COULD_NOT_CALCULATE_IS_COMPLETED": "Σφάλμα: Αδύνατος ο υπολογισμός για το στοιχείο με id",
