+++ src/main/webapp/wise5/i18n/i18n_en.json
@@ -1,6 +1,5 @@
 {
   "ADD_CONNECTED_COMPONENT": "Add Connected Component",
-  "ADD_MILESTONE": "Add Milestone",
   "ADD_TO_NOTEBOOK": "Add To Notebook",
   "ADVANCED": "Advanced",
   "ALERTS": "Alerts",
@@ -22,15 +21,12 @@
   "CORRECT": "Correct",
   "COUNT": "Count",
   "DELETE": "Delete",
-  "DELETE_MILESTONE": "Delete Milestone",
-  "DELETE_MILESTONE_CONFIRM": "Are you sure you want to permanently delete this milestone?",
   "DESCRIPTION": "Description",
   "DETAILS": "Details",
   "DONE": "Done",
   "DOWN": "Down",
   "EDIT": "Edit",
   "EDIT_COMPONENT_RUBRIC": "Edit Component Rubric",
-  "EDIT_MILESTONE": "Edit Milestone",
   "ENTER_PROMPT_HERE": "Enter Prompt Here",
   "ERROR_CHECK_YOUR_INTERNET_CONNECTION": "Error: Data is not being saved! Check your internet connection.",
   "ERROR_COULD_NOT_CALCULATE_IS_COMPLETED": "Error: Could not calculate isCompleted() for component with id ",
